% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_linear_terms.R
\name{add_linear_terms}
\alias{add_linear_terms}
\title{add_linear_terms: Add linear terms to another function.
Allows you to easily change an existing function to include linear terms.}
\usage{
add_linear_terms(func, coeffs)
}
\arguments{
\item{func}{Function to add linear terms to}

\item{coeffs}{Linear coefficients, should have same length as function has dimensions}
}
\value{
Function with added linear terms
}
\description{
add_linear_terms: Add linear terms to another function.
Allows you to easily change an existing function to include linear terms.
}
\examples{
banana(c(.1,.2))
add_linear_terms(banana, coeffs=c(10,1000))(c(.1,.2))
}
