% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_null_dims.R
\name{add_null_dims}
\alias{add_null_dims}
\title{add_null_dims: Add null dimensions to another function.
Allows you to pass in input data with any number of dimensions and it will only keep the first nactive.}
\usage{
add_null_dims(func, nactive)
}
\arguments{
\item{func}{Function to add null dimensions to}

\item{nactive}{Number of active dimensions in func}
}
\value{
Function that can take any dimensional input
}
\description{
add_null_dims: Add null dimensions to another function.
Allows you to pass in input data with any number of dimensions and it will only keep the first nactive.
}
\examples{
banana(c(.1,.2))
# banana(c(.1,.2,.4,.5,.6,.7,.8)) # gives warning
add_null_dims(banana, nact=2)(c(.1,.2,.4,.5,.6,.7,.8))
}
