% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcprofile.R
\name{funcprofile}
\alias{funcprofile}
\title{Profile a function}
\usage{
funcprofile(func, d, n = 1000 * d, bins = 30)
}
\arguments{
\item{func}{A function with a single output}

\item{d}{The number of input dimensions for the function}

\item{n}{The number of points to use for the linear model.}

\item{bins}{Number of bins in histogram.}
}
\value{
Nothing, prints and plots
}
\description{
Gives details about how linear it is.
}
\examples{
funcprofile(ackley, 2)
}
