% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFF.R
\name{nsin}
\alias{nsin}
\alias{vsin}
\title{Wave functions}
\usage{
nsin(xx)

vsin(xx)
}
\arguments{
\item{xx}{Input values}
}
\value{
nsin evaluated at nsin
}
\description{
nsin: Block wave

vsin: v wave
}
\examples{
curve(nsin(2*pi*x), n = 1000)
curve(nsin(12*pi*x), n = 1000)
curve(vsin(2*pi*x), n = 1000)
curve(vsin(12*pi*x), n = 1000)
}
