% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqplots.r
\name{PlotVars}
\alias{PlotVars}
\title{A function to make plot by two variables}
\usage{
PlotVars(data, x, y)
}
\arguments{
\item{data}{dataframe, which should include a StationCode column for your station}

\item{y}{Environmental parameter}
}
\description{
This function makes plots of water quality
}
\examples{
PlotVars(YBFMP_WQ, Conductivity, WaterTemperature)

}
\keyword{QAQC}
