% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastrigin.R
\name{rastrigin}
\alias{rastrigin}
\title{Computes Rastrigin function for n dimensional vector}
\usage{
rastrigin(X, A = 10, n = 2)
}
\arguments{
\item{X}{A numeric vector}

\item{A}{A number.}

\item{n}{Number of dimensions.}
}
\value{
A number
}
\description{
Computes Rastrigin function for n dimensional vector
}
\examples{
rastrigin(c(1, 1))
rastrigin(c(0,0,3), A = 11, n = 3 )
}
