% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFF.R
\name{RFF}
\alias{RFF}
\title{Evaluate an RFF (random wave function) at given input}
\usage{
RFF(x, freq, mag, dirr, offset, wave = sin)
}
\arguments{
\item{x}{Matrix whose rows are points to evaluate}

\item{freq}{Vector of wave frequencies}

\item{mag}{Vector of wave magnitudes}

\item{dirr}{Matrix of wave directions}

\item{offset}{Vector of wave offsets}

\item{wave}{Type of wave}
}
\value{
Output of RFF evaluated at x
}
\description{
Evaluate an RFF (random wave function) at given input
}
\examples{
curve(RFF(x,3,1,1,0))
curve(RFF(x,c(3,20),c(1,.1),c(1,1),c(0,0)), n=1e3)
}

