% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFF.R
\name{RFF_get}
\alias{RFF_get}
\title{Create a new RFF function}
\usage{
RFF_get(D = 2, M = 30, wave = sin)
}
\arguments{
\item{D}{Number of dimensions}

\item{M}{Number of random waves}

\item{wave}{Type of wave}
}
\value{
A random wave function
}
\description{
Create a new RFF function
}
\examples{
func <- RFF_get(D=1)
curve(func)
}

