% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenbrock.R
\name{rosenbrock}
\alias{rosenbrock}
\title{Computes Rosenbrock function for n dimensional vector}
\usage{
rosenbrock(X, n = 2, a = 1, b = 100)
}
\arguments{
\item{X}{A numeric vector}

\item{n}{Number of dimensions.}

\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number
}
\description{
Computes Rosenbrock function for n dimensional vector
}
\examples{
rosenbrock(c(1, 1))
rosenbrock(c(0,0,3), a = 2, b =110, n = 3 )
}
