% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{standard_test_func}
\alias{standard_test_func}
\title{Create a standard test function. This makes it easier to create
many functions that follow the same template.}
\usage{
standard_test_func(func, scale_it_ = F, scale_low_ = NULL,
  scale_high_ = NULL, noise_ = 0, ...)
}
\arguments{
\item{func}{A function that takes a vector representing a single point.}

\item{scale_it_}{Should the function scale the inputs from [0, 1]^D to
[scale_low_, scale_high_] by default? This can be overridden when
actually giving the output function points to evaluate.}

\item{scale_low_}{What is the default lower bound of the data?}

\item{scale_high_}{What is the default upper bound of the data?}

\item{noise_}{Should noise be added to the function by default?}

\item{...}{Parameters passed to func when evaluating points.}
}
\value{
A test function created using the standard_test_func template.
}
\description{
Create a standard test function. This makes it easier to create
many functions that follow the same template.
}
\examples{
gaussian1 <- standard_test_func(.gaussian1, scale_it=F, scale_low = c(0,0), scale_high = c(1,1))
.gaussian1 <- function(x, center=.5, s2=.01) {
  exp(-sum((x-center)^2/2/s2))
}
}

