% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styb_tang.R
\name{styb_tang}
\alias{styb_tang}
\title{Computes Styblinski-Tang function for n dimensional vector}
\usage{
styb_tang(X)
}
\arguments{
\item{X}{A numeric vector}
}
\value{
A number
}
\description{
Computes Styblinski-Tang function for n dimensional vector
}
\examples{
styb_tang(c(1, 1))
styb_tang(c(0,0,3))
}
