% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{test_func_apply}
\alias{OTL_Circuit}
\alias{banana}
\alias{borehole}
\alias{branin}
\alias{currin1991}
\alias{franke}
\alias{gaussian1}
\alias{lim2002}
\alias{powsin}
\alias{sinumoid}
\alias{sqrtsin}
\alias{test_func_apply}
\alias{waterfall}
\alias{zhou1998}
\title{General function for evaluating a test function}
\usage{
test_func_apply(func, x, scale_it, scale_low, scale_high, noise = 0, ...)

branin(x, scale_it = T, scale_low = c(-5, 0), scale_high = c(10, 15))

borehole(x, scale_it = T, scale_low = c(0.05, 100, 63070, 990, 63.1, 700,
  1120, 9855), scale_high = c(0.15, 50000, 115600, 1110, 116, 820, 1680,
  12045))

franke(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

zhou1998(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

currin1991(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

lim2002(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

banana(x, scale_it = T, scale_low = c(-20, -10), scale_high = c(20, 5))

gaussian1(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

sinumoid(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

waterfall(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

sqrtsin(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1))

powsin(x, scale_it = F, scale_low = c(0, 0), scale_high = c(1, 1),
  noise = 0, pow = 1)

OTL_Circuit(x, scale_it = T, scale_low = c(50, 25, 0.5, 1.2, 0.25, 50),
  scale_high = c(150, 70, 3, 2.5, 1.2, 300))
}
\arguments{
\item{func}{A function to evaluate}

\item{x}{Input value}

\item{scale_it}{Should the data be scaled from [0, 1]^D to
[scale_low, scale_high]? This means the input data is confined
to be in [0, 1]^D, but the function isn't.}

\item{scale_low}{Lower bound for each variable}

\item{scale_high}{Upper bound for each variable}

\item{noise}{If white noise should be added, specify the
standard deviation for normal noise}

\item{...}{Additional parameters for func}

\item{pow}{Power for powsin}
}
\value{
Function values at x
}
\description{
General function for evaluating a test function

A function.
2 dimensional function.

A function estimating water flow through a borehole.
8 dimensional function.

A function.
2 dimensional function.

A function.
2 dimensional function.

A function.
2 dimensional function.

Some function?
2 dimensional function.

A banana shaped function.
2 dimensional function.

A Gaussian function centered at 0.5.
Any dimensional function.

A sinusoid added to a sigmoid function.
2 dimensional function.

A sinusoid added to a sigmoid function.
2 dimensional function.

A square root of a sine function.
Any dimensional function.

A sine function raised to a power keeping its original sign.
Any dimensional function.

OTL Circuit.
6 dimensional function.
}
\examples{
x <- matrix(seq(0,1,length.out=10), ncol=1)
y <- test_func_apply(sin, x, TRUE, 0, 2*pi, .05)
plot(x,y)
curve(sin(2*pi*x), col=2, add=TRUE)
branin(runif(2))
borehole(runif(8))
franke(runif(2))
zhou1998(runif(2))
currin1991(runif(2))
lim2002(runif(2))
banana(runif(2))
gaussian1(runif(2))
sinumoid(runif(2))
waterfall(runif(2))
sqrtsin(runif(1))
powsin(runif(1))#,pow=2)
OTL_Circuit(runif(6))
}

