% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions-package.R
\docType{package}
\name{testfunctions-package}
\alias{testfunctions}
\alias{testfunctions-package}
\title{testfunctions: Make water quality plots, just testing package development}
\description{
Testing how to make a package. Some wq plotting code..
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/catarfish/testfunctions}
  \item Report bugs at \url{https://github.com/catarfish/testfunctions/issues}
}

}
\author{
\strong{Maintainer}: catarina Pien \email{catarina.pien@water.ca.gov} (\href{https://orcid.org/0000-0003-4427-6300}{ORCID})

}
\keyword{internal}
