% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_ackley}
\alias{TF_ackley}
\title{TF_ackley: Ackley function for evaluating a single point.}
\usage{
TF_ackley(x, a = 20, b = 0.2, c = 2 * pi)
}
\arguments{
\item{x}{Input vector at which to evaluate.}

\item{a}{A constant for ackley()}

\item{b}{A constant for ackley()}

\item{c}{A constant for ackley()}
}
\value{
Function output evaluated at x.
}
\description{
TF_ackley: Ackley function for evaluating a single point.
}
\examples{
TF_ackley(c(0, 0)) # minimum of zero, hard to solve
}
