% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_bananagramacy2Dexp.R
\name{TF_bananatimesgramacy2Dexp}
\alias{TF_bananatimesgramacy2Dexp}
\title{TF_bananatimesgramacy2Dexp: bananatimesgramacy2Dexp function for evaluating a single point.}
\usage{
TF_bananatimesgramacy2Dexp(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_bananatimesgramacy2Dexp: bananatimesgramacy2Dexp function for evaluating a single point.
}
\examples{
TF_bananatimesgramacy2Dexp(rep(0,6))
TF_bananatimesgramacy2Dexp(rep(1,6))
}
