% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_beale}
\alias{TF_beale}
\title{TF_beale: Beale function for evaluating a single point.}
\usage{
TF_beale(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_beale: Beale function for evaluating a single point.
}
\examples{
TF_beale(rep(0,2))
TF_beale(rep(1,2))
}
