% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_beambending}
\alias{TF_beambending}
\title{TF_beambending: beambending function for evaluating a single point.}
\usage{
TF_beambending(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_beambending: beambending function for evaluating a single point.
}
\examples{
TF_beambending(rep(0,3))
TF_beambending(rep(1,3))
}
