% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_chengsandu}
\alias{TF_chengsandu}
\title{TF_chengsandu: chengsandu function for evaluating a single point.}
\usage{
TF_chengsandu(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_chengsandu: chengsandu function for evaluating a single point.
}
\examples{
TF_chengsandu(rep(0,2))
TF_chengsandu(rep(1,2))
}
\references{
Cheng, Haiyan, and Adrian Sandu. "Collocation least-squares polynomial chaos method." In Proceedings of the 2010 Spring Simulation Multiconference, p. 80. Society for Computer Simulation International, 2010.
}
