% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_detpep8d}
\alias{TF_detpep8d}
\title{TF_detpep8d: detpep8d function for evaluating a single point.}
\usage{
TF_detpep8d(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_detpep8d: detpep8d function for evaluating a single point.
}
\examples{
TF_detpep8d(rep(0,2))
TF_detpep8d(rep(1,2))
}
