% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_easom}
\alias{TF_easom}
\title{TF_easom: Easom function for evaluating a single point.}
\usage{
TF_easom(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_easom: Easom function for evaluating a single point.
}
\examples{
TF_easom(rep(0,2))
TF_easom(rep(1,2))
}
