% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_Gfunction}
\alias{TF_Gfunction}
\title{TF_Gfunction: G-function for evaluating a single point.}
\usage{
TF_Gfunction(x, a = (1:length(x) - 1)/2)
}
\arguments{
\item{x}{Input vector at which to evaluate.}

\item{a}{Parameter for Gfunction}
}
\value{
Function output evaluated at x.
}
\description{
TF_Gfunction: G-function for evaluating a single point.
}
\examples{
TF_Gfunction(rep(0,8))
TF_Gfunction(rep(1,8))
}
