% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_GoldsteinPrice}
\alias{TF_GoldsteinPrice}
\title{TF_GoldsteinPrice: Goldstein Price function for evaluating a single point}
\usage{
TF_GoldsteinPrice(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_GoldsteinPrice: Goldstein Price function for evaluating a single point
}
\examples{
TF_GoldsteinPrice(c(0, -1)) # minimum
}
