% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_GoldsteinPriceLog}
\alias{TF_GoldsteinPriceLog}
\title{TF_GoldsteinPrice: Goldstein Price function for evaluating a single point
on a log scale, normalized to have mean 0 and variance 1.}
\usage{
TF_GoldsteinPriceLog(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_GoldsteinPrice: Goldstein Price function for evaluating a single point
on a log scale, normalized to have mean 0 and variance 1.
}
\examples{
TF_GoldsteinPriceLog(c(0, -1)) # minimum
}
