% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_gramacy2Dexp.R
\name{TF_gramacy2Dexp}
\alias{TF_gramacy2Dexp}
\title{TF_gramacy2Dexp: gramacy2Dexp function for evaluating a single point.}
\usage{
TF_gramacy2Dexp(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_gramacy2Dexp: gramacy2Dexp function for evaluating a single point.
}
\examples{
TF_gramacy2Dexp(rep(0,2))
TF_gramacy2Dexp(rep(1,2))
}
\references{
Gramacy, Robert B., and Herbert KH Lee.
"Adaptive design and analysis of supercomputer experiments."
Technometrics 51.2 (2009): 130-145.
}
