% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_gramacy6D.R
\name{TF_gramacy6D}
\alias{TF_gramacy6D}
\title{TF_gramacy6D: gramacy6D function for evaluating a single point.}
\usage{
TF_gramacy6D(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
From Gramacy and Lee (2009).
}
\examples{
TF_gramacy6D(rep(0,6))
TF_gramacy6D(rep(1,6))
}
\references{
Gramacy, Robert B., and Herbert KH Lee.
"Adaptive design and analysis of supercomputer experiments."
Technometrics 51.2 (2009): 130-145.
}
