% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_griewank}
\alias{TF_griewank}
\title{TF_griewank: Griewank function for evaluating a single point.}
\usage{
TF_griewank(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_griewank: Griewank function for evaluating a single point.
}
\examples{
TF_griewank(rep(0,2))
TF_griewank(rep(1,2))
}
