% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_hartmann}
\alias{TF_hartmann}
\title{TF_hartmann: hartmann function for evaluating a single point.}
\usage{
TF_hartmann(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_hartmann: hartmann function for evaluating a single point.
}
\examples{
TF_hartmann(rep(0,6))
TF_hartmann(rep(1,6))
TF_hartmann(c(.20169, .150011, .476874, .275332, .311652, .6573)) # Global minimum of -3.322368
}
