% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_hump}
\alias{TF_hump}
\title{TF_hump: Hump function for evaluating a single point.}
\usage{
TF_hump(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_hump: Hump function for evaluating a single point.
}
\examples{
TF_hump(rep(0,2))
TF_hump(rep(1,2))
}
