% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_levy}
\alias{TF_levy}
\title{TF_levy: Levy function for evaluating a single point.}
\usage{
TF_levy(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_levy: Levy function for evaluating a single point.
}
\examples{
TF_levy(rep(0,2))
TF_levy(rep(1,2))
}
