% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_levytilt}
\alias{TF_levytilt}
\title{TF_levytilt: Levy function with a tilt for evaluating a single point.}
\usage{
TF_levytilt(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_levytilt: Levy function with a tilt for evaluating a single point.
}
\examples{
TF_levytilt(rep(0,2))
TF_levytilt(rep(1,2))
}
