% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_linkletter_nosignal}
\alias{TF_linkletter_nosignal}
\title{TF_linkletter_nosignal: Linkletter (2006) no signal function for evaluating a single point.}
\usage{
TF_linkletter_nosignal(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_linkletter_nosignal: Linkletter (2006) no signal function for evaluating a single point.
}
\examples{
TF_linkletter_nosignal(rep(0,2))
TF_linkletter_nosignal(rep(1,2))
}
