% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_logistic}
\alias{TF_logistic}
\title{TF_logistic: logistic function for evaluating a single point.}
\usage{
TF_logistic(x, offset = 0, scl = 1)
}
\arguments{
\item{x}{Input vector at which to evaluate.}

\item{offset}{Amount it should be offset}

\item{scl}{Scale parameter}
}
\value{
Function output evaluated at x.
}
\description{
TF_logistic: logistic function for evaluating a single point.
}
\examples{
TF_logistic(0)
TF_logistic(1)
}
