% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_logistic15}
\alias{TF_logistic15}
\title{TF_logistic15: logistic15 function for evaluating a single point.
Same as logistic except adjusted to be reasonable from 0 to 1.}
\usage{
TF_logistic15(x, offset = 0.5, scl = 15)
}
\arguments{
\item{x}{Input vector at which to evaluate.}

\item{offset}{Amount it should be offset}

\item{scl}{Scale parameter}
}
\value{
Function output evaluated at x.
}
\description{
TF_logistic15: logistic15 function for evaluating a single point.
Same as logistic except adjusted to be reasonable from 0 to 1.
}
\examples{
TF_logistic15(0)
TF_logistic15(1)
curve(Vectorize(TF_logistic15)(x))
}
