% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_logistic_plateau}
\alias{TF_logistic_plateau}
\title{TF_logistic_plateau: logistic_plateau function for evaluating a single point.}
\usage{
TF_logistic_plateau(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_logistic_plateau: logistic_plateau function for evaluating a single point.
}
\examples{
TF_logistic_plateau(0)
TF_logistic_plateau(.5)
}
