% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_michalewicz}
\alias{TF_michalewicz}
\title{TF_michalewicz: Michalewicz function for evaluating a single point.}
\usage{
TF_michalewicz(x, m = 10)
}
\arguments{
\item{x}{Input vector at which to evaluate.}

\item{m}{Parameter for the michalewicz function}
}
\value{
Function output evaluated at x.
}
\description{
TF_michalewicz: Michalewicz function for evaluating a single point.
}
\examples{
TF_michalewicz(rep(0,2))
TF_michalewicz(rep(1,2))
}
