% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_moon_high}
\alias{TF_moon_high}
\title{TF_moon_high: Moon (2010) high-dimensional function for evaluating a single point.}
\usage{
TF_moon_high(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_moon_high: Moon (2010) high-dimensional function for evaluating a single point.
}
\examples{
TF_moon_high(rep(0,20))
TF_moon_high(rep(1,20))
}
