% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_morris}
\alias{TF_morris}
\title{TF_morris: morris function for evaluating a single point.}
\usage{
TF_morris(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_morris: morris function for evaluating a single point.
}
\examples{
TF_morris(rep(0,20))
TF_morris(rep(1,20))
}
\references{
http://www.abe.ufl.edu/jjones/ABE_5646/2010/Morris.1991%20SA%20paper.pdf
}
