% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_powsin}
\alias{TF_OTL_Circuit}
\alias{TF_powsin}
\title{A function taking in a single vector.
Any dimensional function.
See corresponding function with "TF_" for more details.}
\usage{
TF_powsin(x, freq = 2 * pi, pow = 0.7)

TF_OTL_Circuit(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}

\item{freq}{Wave frequency}

\item{pow}{Power to raise wave to.}
}
\value{
Function output evaluated at x.
}
\description{
A function taking in a single vector.
Any dimensional function.
See corresponding function with "TF_" for more details.

OTL Circuit function for evaluating a single point
}
\examples{
TF_powsin(runif(2))
TF_OTL_Circuit(c(50,25,0.5,1.2,0.25,50))
}

