% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_piston}
\alias{TF_piston}
\title{TF_piston: Piston simulation function for evaluating a single point.}
\usage{
TF_piston(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_piston: Piston simulation function for evaluating a single point.
}
\examples{
TF_piston(c(30,.005,.002,1e3,9e4,290,340)) # minimum of zero, hard to solve
}
