% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_quad_peaks}
\alias{TF_quad_peaks}
\title{TF_quad_peaks: quad_peaks function for evaluating a single point.}
\usage{
TF_quad_peaks(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_quad_peaks: quad_peaks function for evaluating a single point.
}
\examples{
TF_quad_peaks(rep(0,2))
TF_quad_peaks(rep(1,2))
}
