% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_rastrigin}
\alias{TF_rastrigin}
\title{TF_rastrigin: Rastrigin function for evaluating a single point.}
\usage{
TF_rastrigin(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_rastrigin: Rastrigin function for evaluating a single point.
}
\examples{
TF_rastrigin(rep(0,2))
TF_rastrigin(rep(1,2))
}
