% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_robotarm}
\alias{TF_robotarm}
\title{TF_robotarm: Robot arm function for evaluating a single point.}
\usage{
TF_robotarm(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_robotarm: Robot arm function for evaluating a single point.
}
\examples{
TF_robotarm(rep(0,8))
TF_robotarm(rep(1,8))
}
