% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_RoosArnold}
\alias{TF_RoosArnold}
\title{TF_RoosArnold: Roos & Arnold (1963) function for evaluating a single point.}
\usage{
TF_RoosArnold(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_RoosArnold: Roos & Arnold (1963) function for evaluating a single point.
}
\examples{
TF_RoosArnold(rep(0,8))
TF_RoosArnold(rep(1,8))
}
