% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_SWNExpCos}
\alias{TF_SWNExpCos}
\title{TF_SWNExpCos: SWNExpCos function for evaluating a single point.}
\usage{
TF_SWNExpCos(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_SWNExpCos: SWNExpCos function for evaluating a single point.
}
\examples{
TF_SWNExpCos(rep(0,2))
TF_SWNExpCos(rep(1,2))
}
\references{
Santner, T. J., Williams, B. J., & Notz, W. (2003). The Design and Analysis of Computer Experiments. Springer Science & Business Media.
}
