% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_vertigrad}
\alias{TF_vertigrad}
\title{TF_vertigrad: vertigrad function for evaluating a single point.}
\usage{
TF_vertigrad(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_vertigrad: vertigrad function for evaluating a single point.
}
\examples{
TF_vertigrad(rep(0,2))
TF_vertigrad(rep(1,2))
}
