% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_vertigrad_grad}
\alias{TF_vertigrad_grad}
\title{TF_vertigrad_grad: vertigrad_grad function for evaluating a single point.}
\usage{
TF_vertigrad_grad(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_vertigrad_grad: vertigrad_grad function for evaluating a single point.
}
\examples{
TF_vertigrad_grad(rep(0,2))
TF_vertigrad_grad(rep(1,2))
}
\references{
Forrester, A., & Keane, A. (2008). Engineering design via surrogate modelling: a practical guide. John Wiley & Sons.
}
