% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_welch}
\alias{TF_welch}
\title{TF_welch: Welch function for evaluating a single point.}
\usage{
TF_welch(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_welch: Welch function for evaluating a single point.
}
\examples{
TF_welch(rep(0,20)) # minimum of zero, hard to solve
}
