% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{TF_wingweight}
\alias{TF_wingweight}
\title{TF_wingweight: Wing weight function for evaluating a single point.}
\usage{
TF_wingweight(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_wingweight: Wing weight function for evaluating a single point.
}
\examples{
TF_wingweight(c(150,220,6,-10,16,.5,.08,2.5,1700,.025)) # minimum of zero, hard to solve
}
\references{
Forrester, A., & Keane, A. (2008). Engineering design via surrogate modelling: a practical guide. John Wiley & Sons.
}
