% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{TF_winkel}
\alias{TF_winkel}
\title{TF_winkel: winkel function for evaluating a single point.}
\usage{
TF_winkel(x)
}
\arguments{
\item{x}{Input vector at which to evaluate.}
}
\value{
Function output evaluated at x.
}
\description{
TF_winkel: winkel function for evaluating a single point.
}
\examples{
TF_winkel(rep(0,2))
TF_winkel(rep(1,2))
}
\references{
Winkel, Munir A., Jonathan W. Stallings, Curt B. Storlie, and
Brian J. Reich. "Sequential Optimization in Locally Important Dimensions."
arXiv preprint arXiv:1804.10671 (2018).
}
