% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqplots.r
\name{VisHist}
\alias{VisHist}
\title{A function to make histograms by station and parameter}
\usage{
VisHist(data, y, bin)
}
\arguments{
\item{data}{dataframe, which should include a StationCode column for you station}

\item{y}{Environmental parameter}

\item{bin}{Histogram binwidth}
}
\description{
This function makes plots of water quality
}
\examples{
VisHist(YBFMP_WQ, Conductivity, 10)

}
\keyword{QAQC}
