% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqplots.r
\name{Yearbox}
\alias{Yearbox}
\title{A function to make water boxplot by WY}
\usage{
Yearbox(data, y)
}
\arguments{
\item{data}{dataframe, which should include a StationCode column for you station and WY column for water year}

\item{y}{Environmental parameter}
}
\description{
This function makes plots of WQ
}
\examples{
Yearbox(YBFMP_WQ, Conductivity)

}
\keyword{QAQC}
