% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{numHessian}
\alias{numHessian}
\title{Create function calculating the numerical hessian}
\usage{
numHessian(func, ...)
}
\arguments{
\item{func}{Function to get hessian of}

\item{...}{Arguments passed to numDeriv::hessian().}
}
\value{
A hessian function
}
\description{
Create function calculating the numerical hessian
}
\examples{
numHessian(sin)
}
