% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.r
\name{yearbox}
\alias{yearbox}
\title{yearbox}
\usage{
yearbox(data, x, y, fillby)
}
\arguments{
\item{data}{dataframe}

\item{x}{x axis}

\item{y}{y axis}

\item{fillby}{view by}
}
\value{

}
\description{
yearbox
}
\examples{


}
